+++ R/uc-schemas.R
@@ -10,28 +10,28 @@
 #'
 #' @returns List
 #' @export
-db_uc_schemas_list <- function(catalog,
-                               max_results = 1000,
-                               page_token = NULL,
-                               host = db_host(), token = db_token(),
-                               perform_request = TRUE) {
-
+db_uc_schemas_list <- function(
+  catalog,
+  max_results = 1000,
+  page_token = NULL,
+  host = db_host(),
+  token = db_token(),
+  perform_request = TRUE
+) {
   stopifnot(max_results <= 1000)
-
-  body <- list(
-    max_results = max_results,
-    page_token = page_token
-  )
 
   req <- db_request(
     endpoint = "unity-catalog/schemas",
     method = "GET",
     version = "2.1",
     host = host,
-    token = token,
-    body = body
+    token = token
   ) |>
-    httr2::req_url_query(catalog_name = catalog)
+    httr2::req_url_query(
+      catalog_name = catalog,
+      max_results = max_results,
+      page_token = page_token
+    )
 
   if (perform_request) {
     db_perform_request(req)$schemas
@@ -53,11 +53,14 @@ db_uc_schemas_list <- function(catalog,
 #'
 #' @returns List
 #' @export
-db_uc_schemas_get <- function(catalog, schema,
-                              include_browse = TRUE,
-                              host = db_host(), token = db_token(),
-                              perform_request = TRUE) {
-
+db_uc_schemas_get <- function(
+  catalog,
+  schema,
+  include_browse = TRUE,
+  host = db_host(),
+  token = db_token(),
+  perform_request = TRUE
+) {
   body <- list(
     include_browse = include_browse
   )
@@ -78,4 +81,3 @@ db_uc_schemas_get <- function(catalog, schema,
     req
   }
 }
-