+++ tests/testthat/test-uc-catalogs.R
@@ -1,19 +1,26 @@
 test_that("Unity Catalog: Catalogs API - don't perform", {
-
   withr::local_envvar(c(
     "DATABRICKS_HOST" = "http://mock_host",
     "DATABRICKS_TOKEN" = "mock_token"
   ))
 
-  resp_catalog_list <- db_uc_catalogs_list(perform_request = FALSE)
+  resp_catalog_list <- db_uc_catalogs_list(
+    max_results = 50,
+    include_browse = FALSE,
+    page_token = "abc",
+    perform_request = FALSE
+  )
   expect_s3_class(resp_catalog_list, "httr2_request")
+  query <- httr2::url_parse(resp_catalog_list$url)$query
+  expect_identical(query$max_results, "50")
+  expect_identical(query$include_browse, "false")
+  expect_identical(query$page_token, "abc")
 
   resp_catalog_get <- db_uc_catalogs_get(
     catalog = "some_catalog",
     perform_request = FALSE
   )
   expect_s3_class(resp_catalog_get, "httr2_request")
-
 })
 
 skip_on_cran()
@@ -21,7 +28,6 @@ skip_unless_authenticated()
 skip_unless_aws_workspace()
 
 test_that("Unity Catalog: Catalogs API", {
-
   expect_no_error({
     resp_catalog_list <- db_uc_catalogs_list()
   })
@@ -32,5 +38,4 @@ test_that("Unity Catalog: Catalogs API", {
       catalog = resp_catalog_list[[1]]$name
     )
   })
-
 })